add initial version of hypervisor tooling (#18)
* deb: fix extraction of uncompressed data.tar

In some debian packages the data.tar is not compressed
and named only data.tar without any extension.

* add initial version of hypervisor tooling

Adds tooling to generated the hypervisor configuration (lua scripts)
using a yaml configuration file and a yaml model of this configuration
described in yaml and python. The reason for having the model "twice", is that
the yaml definition of the model describes how the configuration is parsed into
a python object model. The python model is required to do some postprocessing
of the configuration, before jinja2 is used to generate the final configuration.

The used hypervisor version can support additional features or require a different
generated configuration. For that reason the configuration model can be extended
by a configuration specialization bundled together with the hypervisor.