debootstrap: add option to not install recommended packages
apt install by default installs all recommends.
This can install a huge number of packages, which are not wanted for
a small embedded system.

This adds an option to not install the recommends.
It should be the default, but to keep the tooling backward compatible,
it defaults to install recommends.