initrd_generator: implement dependencies and module lookup
This implements automatic module dependency resolution and changes the default
format for modules listed in the yaml file from path to just module name.
("kernel/net/bridge/bridge.ko" becomes "bridge").

All modules that are dependencies of the specified modules are
copied to the initrd automatically and due to a correctly generated modules.dep
file, they are loaded automatically as well.

This uses the hosts depmod utility (from kmod), which is not architecture agnostic
so it works also for cross building.