enable parallel execution of cache using tools
The old cache implementation did not handle parallel
execution of tools (e..g initrd and boot generator) very well.
It could corrupt the cache index and even the packages.

This new implementation uses an sqlite database to allow
concurrent read access and synchronized write access to the index.
A package can now only be added once to the index. If a second process
tries to add it, it is just ignored.

Aditionally the cache can be recovered from the  package files stored in the
cache directory if the index does not exist or is corrupted.