development: fix devcontainer setup
The only supported workflow for developing in this repository
should be using the devcontainer:

Update extensions to remove unused ones and add only:
- python
- mypy
- flake8
 - pytest
 - robottest

The python modules are now installed into a venv, because
the global installation in the devcontainer could belong
to a user with a different UID due to the way the devcontainers work.
The local venv is picked up by vscode and used for all vscode extensions and
in the console automatically.