debootstrap: update ca-certificates using the host
The root certificates contained in the ca-certificates pacakges are
requried to download files using https. The initial package comes from
the debootstrap archive. This pacakge may be outdated and updating
it in the target filesystem using apt update && apt upgrade fails, because
the repository may use a certificate, that is not yet part of the ca-certificates package.

This change downloads the latest ca-certifiactes package using the apt proxy on the host
and installs it on the target before runnign apt update.