fake: drop capture_output parameter and close stdin
The main goal here is to not have output written directly to stdout and stderr.

The capture_output parameter does not make a lot of sense.
With capture_output=False the output was written directly to stdout and
with capture_output=True the output was captured and written to the log
and returned from the function.

Now the output (stdout and stderr) is always caputred and returned and
written to the log. I.e. it is not written to stdout anymore.

When an out stream is passed (stdout parameter != None), the output
is not written to the log because it could be binary (i.e. it is used for cpio output).
This is no change from the previous implementation.

Additionally, stdin is set to subprocess.DEVNULL. This prevents processes
executed from changing the terminal settings (i.e. changing the behavior of
\n to just cause newline instead of a new line and carriage return).
Changing the terminal settings can mess up the output during parallel builds.